Game Loop Completion
The game loop should be complete now. There is some test comments that will have to be removed, but other than that everything should work once the card functions are given. The board print out statement is there, but the other half of the board hasn't been printed yet. The pawn locations still need to be printed.